Issue #1: finished first draft of IAR methods
Instead of the earlier MICCAI notation, I stuck with the notation from
my thesis, which I believe is easier to remember.

In terms of diagrams, I only included the graphical model, but it
would be nice to have some diagrams that give an intuition behind some
of the random variables and their distributions.